requirements.txt
construct==2.10.68
construct-typing==0.6.2
fusepy==3.0.1
rich==13.7.1